organize navLinks prop
