+++ frontend/src/App.js
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import Container from 'react-bootstrap/Container';
 import Header from './components/Header';
 import 'bootstrap/dist/css/bootstrap.min.css';
@@ -6,13 +7,33 @@ import Home from './pages/Home';
 import Wallet from './components/Wallet';
 
 function App() {
+    // const [navLinks, setNavLinks] = useState([]);
+    const [walletSide, setWalletSide] = useState(false);
+
+    const navLinks = [
+        { name: 'Wallet', url: '/wallet'},
+        { name: 'Faucet', url: '/faucet'},
+        { name: 'Miner', url: '/miner'},
+    ];
+
+    const walletLinks = [
+        {name: 'Explorer', url: '/'},
+        {name: 'Create', url: '/create-wallet'},
+        {name: 'Open', url: '/open-wallet'},
+    ];
+
+    if (walletSide) navLinks = walletLinks;
+
     return (
         <div className='App'>
             <Container >
-                <Header />
+                <Header navLinks={navLinks} />
                 <Routes>
                     <Route path='/' element={<Home />} />
-                    <Route path='/wallet' element={<Wallet />} />
+                    <Route 
+                        path='/wallet' 
+                        element={<Wallet setWalletSide={setWalletSide} />}
+                    />
                 </Routes>
             </Container>
         </div>
